requirements.txt
xarray~=0.18.2
pandas~=1.5.1
dash~=2.14.2
rasterio~=1.3.9
dash-leaflet
numpy~=1.20.3
plotly~=5.18.0
geopandas~=0.13.2
shapely~=2.0.2
folium~=0.15.0
pillow~=8.3.1
pyproj~=3.3.1
python-dotenv